+++ lib/util.dart
@@ -79,6 +79,8 @@ String getErrorName(int errorCode) {
       return 'WRONG_KEYS_ERROR';
     case INVALID_PREPROCESS_ERROR:
       return 'INVALID_PREPROCESS_ERROR';
+    case INVALID_DERIVATION_ERROR:
+      return 'INVALID_DERIVATION_ERROR';
     case FEE_ERROR:
       return 'FEE_ERROR';
     case INVALID_PARTICIPANTS_AMOUNT_ERROR: